requirements.txt
discord.py==1.7.3
requests
python-dateutil
python-dotenv
discord-py-slash-command==3.0.3
bs4
pytz